Strengthen fastpath (#169)
This change adds a test which probes for issues with sparse inputs on the fast path.

This fixes #163 in multiple ways. Each of the improvements have been tested individually and in various combinations with weakened versions of other parts of the algorithm to insure they work as intended and provide a benefit.

Signed-off-by: Tom Kaitchuck <Tom.Kaitchuck@gmail.com>