Remove build-time check for `stdsimd` feature
This is blocking rust-lang/rust#117372, which replaces `stdsimd` with
more fine-grained features. However the auto-detection in aHash breaks
when bootstrapping Rust because it detects the `stdsimd` feature on the
old toolchain which is not present on the newly build libcore.

This PR removes the build-time detection of the `stdsimd` feature and
instead just uses the ARM AES intrinsics directly since they are now
stable, but only on AArch64.